+++ src/com/lftechnology/java/training/alina/collections/generics/wildcards/CustomerB.java
@@ -8,7 +8,7 @@ package com.lftechnology.java.training.alina.collections.generics.wildcards;
  */
 public class CustomerB extends Customer {
 
-	public CustomerB(String name, String address, float bill) {
+	public CustomerB(String name, String address, int bill) {
 
 		super(name, address, bill);
 	}
